refactor: streamline legacy team setting migration and update starlight bonus logic
