+++ app/base_combination.py
@@ -319,9 +319,8 @@ class MirrorTeamCombination(QFrame):
             return
 
         data: dict = cfg.yaml.load(setting)
-        from module.config.config_typing import TeamSetting, migrate_legacy_team_setting_data
+        from module.config import TeamSetting
 
-        data = migrate_legacy_team_setting_data(data)
         try:
             team_config = TeamSetting(**data)
         except Exception:
